Unique animal number functionality has been added.

+++ Backend/AnimalRescue/AnimalRescue.API/Models/SequenceModel.cs
@@ -8,6 +8,6 @@ namespace AnimalRescue.API.Models.Tags
     {
         [JsonPropertyName(PropertyConstants.Animal.Number)]
         [JsonProperty(PropertyConstants.Animal.Number)]
-        public long Number { get; set; }
+        public int Number { get; set; }
     }
 }
